requirements.txt
assemblyai==0.26.0
customtkinter==5.2.2
moviepy==2.0.0.dev2
nltk==3.8.1
numpy==1.26.4
Pillow==10.3.0
pygame==2.5.2
pysrt==1.1.2
scikit_learn==1.4.2
textblob==0.18.0.post0
tkvideoplayer==2.8
